driver: use configurable io routines in HW and ensure that driver loops to solve all IRQs.
The loop is mandatory when edge triggered interrupts are used.
It is necessary to achieve state when all enabled interrupts are
served and cleared or disabled and actual IRQ pin or MSI
request is inactive.

Some more cleanup to access chip only through routines
and register description found in ctu_can_fd_hw.h
and ctu_can_fd_regs.h.

Signed-off-by: Pavel Pisa <pisa@cmp.felk.cvut.cz>