src: Add DFF to Bit Destuffing processing.
Enable CRC in Idle only after DOMINANT was sampled in Sample point,
not during whole time when RX Data are dominant! Previous value is
sampled anyway in bit destuffing and since it is first dominant,
same_bits counter needs to be set to 1 anyway! So it will be set
to 1 by rising edge on enable after DOMINANT is sampled and we are
sure we go to BASE_ID.