src: Handle simultaneous Set Abort and Unlock, Arb. lost, Error.
If TXT Buffer is in TX In progress when Unlock arbitration lost,
or Error frame occurs and retransmitt limit is not reached TXT Buffer
goes to Ready.
IF simultaneously SW Abort command comes, HW command should not have
priority, and TXT Buffer shall be aborted (not moved to Ready).
If Buffer went to Ready, then Abort command would be effectively lost
because upon next Lock, it would go directly to TX in progress...