src: throw away Bit-time FSM transition condition.
This should be obsolete as transition is only available from
tseg2 (when in tseg1, h_sync does not cause state transition).
Also, from reset, there should be no hard-sync.

And if there is transition from tseg2 to tseg1 due to hsync,
this is already included in segm_end.

Unit Test passing.