src: Mask TXT Buffer output.
Output is masked always when TXT Buffer FSM is not accessed by TX
Arbitrator or CAN Core. This removes lot of warning logs on
arithmetic operations on undefined values! Also design is safer
because, it makes sure that RAM truly is not read by CAN Core nor
Arbitrator when it should not be! On the other hand, it costs slightly
more logic...