src: IDE in arbitration not in control for receiver!
When we don't transmitt anything, keep IDE in Arbitration field,
not in Control field! This is better because we won't receive
short is_control pulse for the duration of IDE bit!! We don't mind
that in case of Error, this will be always interpreted as error in
Arbitration field even if BASE frame is being received!