test: Debug error rules E feature test.
If error is forced on 5th bit of error delimiter, this will be already
after error frame of the other node, so next error frame will hit Error
delimiter of the other node. Due to this, the other node will transmit
next error frame and thus after bus is idle, REC will be incremented
twice. This is not a bug, but check of REC must be done directly after
error is sampled!