test: Debug form error in Control field feature test.
Remove r0 on CAN 2.0 extended frame format! This should NOT be
treated as form error when Recessive is detected!

Swap test to check on receiver as on transmitter this is now
detected as bit error, not form error. Selectively enable/disable
FD support so that form error can be detected at correct spot!