driver: correct formatting according to 5.4 Linux kernel checkpatch reports.
Some lines are longer than 80 characters to conform alignment rules
of follow-up lines in function definitions/declarations requested
by checkpatch. 80 characters limit has been relaxed to 100 in 5.7
kernel so this should be without problems for mainline inclusion.

Suggestion

  CHECK: Prefer kernel type 'u32' over 'uint32_t'

ignored in auto-generated header files. The files are used in QEMU
and user space tools and correct setup of u32 definition to not
clash in all cases could be problematic.

Signed-off-by: Pavel Pisa <pisa@cmp.felk.cvut.cz>