Update ctucanfd_base.c
According to the datasheet (see section 3.1.46 CTU_CANFD_Datasheet.pdf), Secondary Sampling Point is not used when SSP_SRC = 0b01. This is true when the bit rate is less than or equal to 1 Mbit/s. But if the CAN FD interface works with a higher data bit rate, for example 4 or 5 Mbit/s, then the transceiver delay value becomes significant and you need to use
SSP_SRC = 0x0 or 0x2.